+++ projects/swimlane/ngx-ui/src/lib/components/tip/tip.component.ts
@@ -1,6 +1,16 @@
 import { Component, Input, Output, EventEmitter, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
 import { TipStatus } from './tip-status.enum';
 
+const ICONS = {
+  [TipStatus.Error]: 'warning-filled-sm',
+  [TipStatus.Warning]: 'alert',
+  default: 'info-filled-small'
+};
+
+function getIcon(status: TipStatus): string {
+  return ICONS[status] || ICONS['default'];
+}
+
 @Component({
   selector: 'ngx-tip',
   exportAs: 'ngxTip',
@@ -12,7 +22,8 @@ import { TipStatus } from './tip-status.enum';
     class: 'ngx-tip',
     '[class.ngx-tip--success]': 'status === TipStatus.Success',
     '[class.ngx-tip--error]': 'status === TipStatus.Error',
-    '[class.ngx-tip--notice]': 'status === TipStatus.Notice'
+    '[class.ngx-tip--notice]': 'status === TipStatus.Notice',
+    '[class.ngx-tip--warning]': 'status === TipStatus.Warning'
   }
 })
 export class TipComponent {
@@ -20,13 +31,16 @@ export class TipComponent {
   status: TipStatus;
   @Input()
   isCloseable: boolean = false;
+  @Input()
+  icon: string;
   @Output()
   close = new EventEmitter();
-  icon: string;
   readonly TipStatus = TipStatus;
 
-  ngOnInit() {
-    this.icon = this.status === TipStatus.Error ? 'warning-filled-sm' : 'info-filled-small';
+  ngOnChanges() {
+    if (!this.icon) {
+      this.icon = getIcon(this.status);
+    }
   }
 
   ngOnDestroy() {