+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+- Enhancement: Add `warning` status to `tip` component
+- Enhancement: Add ability to display custom icon on `tip` component
 - Enhancement: Add ability to add new values via `Shift + ⏎` on select component when `allowAdditions` is true
 
 ## 35.6.4 (2021-05-25)